fix: add entity
